% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.fts_ts}
\alias{autoplot.fts_ts}
\title{Plot time series of fts basis coefficients}
\usage{
\method{autoplot}{fts_ts}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fts_ts} containing time-varying
basis function coefficients extracted from an \code{ffc_gam} object}

\item{...}{Ignored}
}
\description{
Produces a time series plot of basis function coefficients from \code{fts_ts} objects
}
\author{
Nicholas J Clark
}
