% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{autoplot}
\alias{autoplot}
\title{Automatic plotting}
\usage{
autoplot(object, ...)
}
\description{
Automatic plotting method for ffc objects. In ffc, used for plotting
time-varying coefficients from \code{\link[=fts_coefs]{fts_coefs()}}.
}
\examples{
# coefs <- fts_coefs(model, summary = FALSE)
# autoplot(coefs)
}
\seealso{
\code{\link[ggplot2]{autoplot}}, \code{\link[=autoplot.fts_ts]{autoplot.fts_ts()}}
}
