% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{bam}
\alias{bam}
\title{Large dataset GAM fitting}
\usage{
bam(
  formula,
  family = gaussian(),
  data = list(),
  weights = NULL,
  subset = NULL,
  na.action = na.omit,
  offset = NULL,
  method = "fREML",
  control = list(),
  select = FALSE,
  scale = 0,
  gamma = 1,
  knots = NULL,
  sp = NULL,
  min.sp = NULL,
  paraPen = NULL,
  chunk.size = 10000,
  rho = 0,
  AR.start = NULL,
  discrete = FALSE,
  cluster = NULL,
  nthreads = 1,
  gc.level = 0,
  use.chol = FALSE,
  samfrac = 1,
  coef = NULL,
  drop.unused.levels = TRUE,
  G = NULL,
  fit = TRUE,
  drop.intercept = NULL,
  in.out = NULL,
  ...
)
}
\description{
Fits GAMs for large datasets using mgcv::bam. In ffc, specify
\code{engine = "bam"} in \code{\link[=ffc_gam]{ffc_gam()}} for large datasets.
}
\examples{
# Use bam engine for large datasets
# mod <- ffc_gam(y ~ fts(x), data = large_data, time = "time", engine = "bam")
}
\seealso{
\code{\link[mgcv]{bam}}, \code{\link[=ffc_gam]{ffc_gam()}}
}
