% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{betar}
\alias{betar}
\title{Beta regression family}
\usage{
betar(theta = NULL, link = "logit", eps = .Machine$double.eps * 100)
}
\description{
Beta distribution family for data on (0,1) interval.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument.
}
\examples{
# ffc_gam(proportions ~ fts(x), data = data, family = betar(), time = "time")
}
\seealso{
\code{\link[mgcv]{betar}}
}
\keyword{internal}
