% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{cnorm}
\alias{cnorm}
\title{Censored normal family}
\usage{
cnorm(theta = NULL, link = "identity")
}
\description{
Censored normal distribution family for censored continuous data.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument.
}
\examples{
# ffc_gam(censored_y ~ fts(x), data = data, family = cnorm(), time = "time")
}
\seealso{
\code{\link[mgcv]{cnorm}}
}
