% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc-package.R
\docType{package}
\name{ffc-package}
\alias{ffc}
\alias{ffc-package}
\title{ffc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
