% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc_gam.R
\name{ffc_gam}
\alias{ffc_gam}
\title{Fit a functional time series model using dynamic functional coefficients}
\usage{
ffc_gam(
  formula,
  family = gaussian(),
  data = list(),
  time,
  engine = c("gam", "bam"),
  ...
)
}
\arguments{
\item{formula}{A GAM formula (see \code{\link[mgcv]{formula.gam}} and also
\code{\link[mgcv]{gam.models}}). This is exactly like the formula for a
GLM except that smooth terms, \code{\link[=fts]{fts()}}, \code{\link[mgcv]{s}()} and
\code{\link[mgcv]{te}()} can be added to the right hand side to specify
that the linear predictor depends on smooth functions of predictors
(or linear functionals of these).}

\item{family}{
This is a family object specifying the distribution and link to use in
fitting etc. See \code{\link{glm}} and \code{\link{family}} for more
details. The extended families listed in \code{\link[mgcv]{family.mgcv}} can also be used.
}

\item{data}{ A data frame or list containing the model response variable and 
covariates required by the formula. By default the variables are taken 
from \code{environment(formula)}: typically the environment from 
which \code{gam} is called.}

\item{time}{\code{character} specifying which variable in \code{data} represents the
the time ordering of the observations}

\item{engine}{\code{character} string specifying which \pkg{mgcv} interface to use
for fitting the model.}

\item{...}{other arguments to pass to either \code{\link[mgcv]{gam}}}
}
\value{
An object of class \code{ffc_gam}, which inherits from objects of class \code{gam} or
\code{bam}
}
\description{
Fit Generalized Additive Models that can include time-varying (dynamic)
functions
}
\details{
This function will update the supplied \code{formula} to ensure any time-varying
functionals (supplied through \code{\link[=fts]{fts()}} terms in the formula right hand side) are
appropriately incorporated into the model. It then passes the updated model and data
objects to the specified \code{engine} for model fitting
}
\examples{
# Fit a dynamic function-on-function regression to the Queensland
# mortality data
data("qld_mortality")
mod <- ffc_gam(
  deaths ~
    offset(log(population)) +
    sex +
    fts(age,
      k = 8, bs = "cr",
      time_bs = "cr", time_k = 10
    ),
  time = "year",
  data = qld_mortality,
  family = poisson(),
  engine = "bam"
)
class(mod)
summary(mod)

# Predictions work in the usual way
head(predict(mod, type = "link"))
head(predict(mod, type = "response"))

# Extract basis coefficient time series
(functional_ts <- fts_coefs(mod))

# Binomial model with cbind() response for trials data
data("qld_mortality")
# Create binomial version of the data for demonstration
qld_binomial <- qld_mortality
qld_binomial$trials <- qld_binomial$population
qld_binomial$successes <- qld_binomial$deaths

mod_binomial <- ffc_gam(
  cbind(successes, trials - successes) ~
    sex +
    fts(age,
      k = 6, bs = "cr",
      time_bs = "cr", time_k = 8
    ),
  time = "year",
  data = qld_binomial,
  family = binomial(),
  engine = "bam"
)
summary(mod_binomial)

}
\seealso{
\code{\link[=fts]{fts()}}, \code{\link[mgcv]{gam}}, \code{\link[mgcv]{bam}}
}
\author{
Nicholas J Clark
}
