% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.fts_ts}
\alias{forecast.fts_ts}
\title{Forecasting functional basis coefficients}
\usage{
\method{forecast}{fts_ts}(object, model = "ARIMA", h = 1, times = 25, ...)
}
\arguments{
\item{object}{An object of class \code{fts_ts} containing time-varying
basis function coefficients extracted from an \code{ffc_gam} object}

\item{model}{A \code{character} string representing a valid univariate model definition
from the \pkg{fable} package. Note that the chosen method must have an associated
\code{generate()} method in order to simulate forecast realisations. Valid models
currently include: \code{'ETS'}, \code{'ARIMA'}, \code{'AR'}, \code{'NAIVE'}, and \code{'NNETAR'}}

\item{h}{A positive \code{integer} specifying the length of the forecast
horizon}

\item{times}{A positive \code{integer} specifying the number of forecast
realisation paths to simulate from the fitted forecast \code{model}}

\item{...}{Ignored}
}
\value{
A \code{tsibble} object containing the forecast prediction (\code{.sim}) for each
replicate realisation (\code{.rep}) at each timestep in the forecast horizon \code{h}
}
\description{
Forecasting functional basis coefficients
}
\details{
Basis function coefficient time series will be used as input
to the specified \code{model} to train a forecasting model that will then be
extrapolated \code{h} timesteps into the future. A total of \code{times} forecast realisations
will be returned for each basis coefficient
}
\seealso{
\code{\link[=predict]{predict()}}, \code{\link[=fts]{fts()}}
}
\author{
Nicholas J Clark
}
