% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{forecast}
\alias{forecast}
\title{Forecasting methods}
\usage{
forecast(object, ...)
}
\description{
Generic forecasting function. In ffc, use with \code{\link[=forecast.fts_ts]{forecast.fts_ts()}} and
\code{\link[=forecast.ffc_gam]{forecast.ffc_gam()}} for functional forecasting.
}
\examples{
# coefs <- fts_coefs(model)
# forecast(coefs, h = 5, model = "ARIMA")
}
\seealso{
\code{\link[generics]{forecast}}, \code{\link[=forecast.fts_ts]{forecast.fts_ts()}}, \code{\link[=forecast.ffc_gam]{forecast.ffc_gam()}}
}
