% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts.R
\name{fts}
\alias{fts}
\title{Define functions with dynamic coefficients in \pkg{ffc} formulae}
\usage{
fts(
  ...,
  mean_only = FALSE,
  k = NA,
  time_k = 10,
  bs = "cr",
  time_bs = "cr",
  m = NA,
  time_m = 2,
  d = NA,
  by = NA,
  xt = NULL,
  pc = NULL
)
}
\arguments{
\item{...}{a list of variables that are the covariates that this smooth is a function of.
Transformations whose form depends on the values of the data are best avoided here:
e.g. \code{fts(log(x), z)} is fine, but \code{fts(I(x / sd(x)), z)} is not.}

\item{mean_only}{\code{Logical} indicating whether to only include a single basis function
to be modelled as time-varying. This can be helpful if you wish to include a temporal
trend that can then be forecasted ahead using appropriate time series models.
Default is \code{FALSE}}

\item{k}{the dimension(s) of the bases used to represent the smooth term.
If not supplied then set either to \code{10} (if only a single covariate is supplied) or
to \code{5 ^ d}, where \code{d} is the number of covariates supplied.
If supplied as a single number then this basis dimension is used for each basis.
If supplied as an array then the elements are the dimensions of the component (marginal)
bases of the tensor product.}

\item{time_k}{the dimension of the bases to be used in the time-varying coefficient
smooths (see \strong{Details} below). Arbitrarily set to \code{10} by default.}

\item{bs}{array (or single character string) specifying the type for each 
marginal basis. \code{"cr"} for cubic regression spline; \code{"cs"} for cubic
regression spline with shrinkage; \code{"cc"} for periodic/cyclic 
cubic regression spline; \code{"tp"} for thin plate regression spline;
\code{"ts"} for t.p.r.s. with extra shrinkage. See \code{\link[mgcv]{smooth.terms}} for details 
and full list. User defined bases can 
also be used here (see \code{\link[mgcv]{smooth.construct}} for an example). If only one 
basis code is given then this is used for all bases.}

\item{time_bs}{a two letter \code{character} string indicating the (penalized)
smoothing basis to use for the time-varying basis coefficients (eg \code{"tp"} for thin
plate regression spline, \code{"cr"} for cubic regression spline).
see \code{\link[mgcv]{smooth.terms}} for an over view of what is available.}

\item{m}{The order of the spline and its penalty (for smooth classes that use this) for each term. 
   If a single number is given  then it is used for all terms. A vector can be used to 
  supply a different \code{m} for each margin. For marginals that take vector \code{m} 
  (e.g. \code{\link[mgcv]{p.spline}} and \code{\link[mgcv]{Duchon.spline}}), then
  a list can be supplied, with a vector element for each margin. \code{NA} autoinitializes. 
  \code{m} is ignored by some bases (e.g. \code{"cr"}).}

\item{time_m}{the order of the penalty for the time-varying coefficient smooths
(e.g. \code{2} for normal cubic spline penalty with 2nd derivatives).
Only some smooth classes use this. The \code{"ps"} class can use a \code{2} item \code{array}
giving the basis and penalty order separately.}

\item{d}{array of marginal basis dimensions. For example if you want a smooth for 3 covariates 
made up of a tensor product of a 2 dimensional t.p.r.s. basis and a 1-dimensional basis, then 
set \code{d=c(2,1)}. Incompatibilities between built in basis types and dimension will be
resolved by resetting the basis type.}

\item{by}{a \code{factor} variable of the same dimension as each covariate, used to create
a replicate of the smooth for each factor level.}

\item{xt}{Either a single object, providing any extra information to be passed
to each marginal basis constructor, or a list of such objects, one for each
marginal basis. }

\item{pc}{If not \code{NULL}, signals a point constraint: the smooth should pass through zero at the
point given here (as a vector or list with names corresponding to the smooth names). Never ignored
if supplied. See \code{\link[mgcv]{identifiability}}. }
}
\value{
A class \code{fts.spec} object defining the smooth function
to be evaluated and configured for estimating time-varying basis function coefficients
}
\description{
Set up smooth terms with time-varying (dynamic) coefficients for use in
\pkg{ffc} models
}
\details{
\pkg{ffc} will evaluate the basis from these smooths and add the basis functions
to the internal model design matrix. Once these basis function predictors are added to the data,
a model will then be estimated that allows their coefficients to change through time using
terms such as
\code{s(time, by = bfun_1, id = 1, bs = 'cr') + s(time, by = bfun_2, id = 1, bs = 'cr') + ...}.
By linking the smoothing parameters using the \code{id} argument, the time-varying function will
be efficiently regularised.
}
\author{
Nicholas J Clark
}
