% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{gam}
\alias{gam}
\title{Generalized Additive Model fitting}
\usage{
gam(
  formula,
  family = gaussian(),
  data = list(),
  weights = NULL,
  subset = NULL,
  na.action,
  offset = NULL,
  method = "GCV.Cp",
  optimizer = c("outer", "newton"),
  control = list(),
  scale = 0,
  select = FALSE,
  knots = NULL,
  sp = NULL,
  min.sp = NULL,
  H = NULL,
  gamma = 1,
  fit = TRUE,
  paraPen = NULL,
  G = NULL,
  in.out = NULL,
  drop.unused.levels = TRUE,
  drop.intercept = NULL,
  nei = NULL,
  discrete = FALSE,
  ...
)
}
\description{
Fits GAMs using mgcv. In ffc, use \code{\link[=ffc_gam]{ffc_gam()}} instead which extends
gam functionality with time-varying coefficients.
}
\examples{
# Use ffc_gam() for time-varying functionality
# mod <- ffc_gam(y ~ fts(x), data = data, time = "time")
}
\seealso{
\code{\link[mgcv]{gam}}, \code{\link[=ffc_gam]{ffc_gam()}}
}
