% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{generate}
\alias{generate}
\title{Generate method}
\usage{
generate(x, ...)
}
\description{
Generic generate function from the generics package.
Used internally for forecast simulation.
}
\examples{
# Used internally by forecast methods
}
\seealso{
\code{\link[generics]{generate}}
}
\keyword{internal}
