% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_data.R
\docType{data}
\name{growth_data}
\alias{growth_data}
\title{Berkeley growth study data}
\format{
A \code{data.frame} containing the following fields:
\describe{
\item{height_cm}{numeric height, in centimeters}
\item{age_yr}{numeric age at which the boys were measured, in years}
\item{id}{factor indicating the different subjects}
}
}
\usage{
growth_data
}
\description{
A dataset containing the heights of 39 boys measured at ages ranging from 1 to 18 years
}
\references{
Tuddenham, R. D., and Snyder, M. M. (1954) "Physical growth of
California boys and girls from birth to age 18", \emph{University of
California Publications in Child Development}, 1, 183-364.
}
\keyword{datasets}
