% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{ldTweedie}
\alias{ldTweedie}
\title{Tweedie log density}
\usage{
ldTweedie(
  y,
  mu = y,
  p = 1.5,
  phi = 1,
  rho = NA,
  theta = NA,
  a = 1.001,
  b = 1.999,
  all.derivs = FALSE
)
}
\description{
Computes log density for Tweedie distribution.
Utility function for Tweedie family.
}
\examples{
# ldTweedie(y = data, mu = fitted_values, p = 1.5, phi = 1)
}
\seealso{
\code{\link[mgcv]{ldTweedie}}, \code{\link[=tw]{tw()}}
}
\keyword{internal}
