% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.frame.R
\name{model.frame.ffc_gam}
\alias{model.frame.ffc_gam}
\title{Extract model.frame from a fitted \code{ffc_gam} object}
\usage{
\method{model.frame}{ffc_gam}(formula, ...)
}
\arguments{
\item{formula}{a model \code{\link[stats]{formula}} or \code{\link[stats]{terms}}
    object or an \R object.}

\item{...}{Ignored}
}
\value{
A \code{matrix} containing the fitted model frame
}
\description{
Extract model.frame from a fitted \code{ffc_gam} object
}
\author{
Nicholas J Clark
}
