% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{multinom}
\alias{multinom}
\title{Multinomial family}
\usage{
multinom(K = 1)
}
\description{
Multinomial distribution family for categorical response data.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument.
}
\examples{
# ffc_gam(categories ~ fts(x), data = data, family = multinom(), time = "time")
}
\seealso{
\code{\link[mgcv]{multinom}}
}
\keyword{internal}
