% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{nb}
\alias{nb}
\title{Negative binomial family}
\usage{
nb(theta = NULL, link = "log")
}
\description{
Negative binomial distribution family for count data with overdispersion.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument.
}
\examples{
# ffc_gam(counts ~ fts(x), data = data, family = nb(), time = "time")
}
\seealso{
\code{\link[mgcv]{nb}}
}
\keyword{internal}
