% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{negbin}
\alias{negbin}
\title{Negative binomial family (alternative)}
\usage{
negbin(theta = stop("'theta' must be specified"), link = "log")
}
\description{
Alternative negative binomial family specification.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument.
}
\examples{
# ffc_gam(counts ~ fts(x), data = data, family = negbin(), time = "time")
}
\seealso{
\code{\link[mgcv]{negbin}}
}
\keyword{internal}
