% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{ocat}
\alias{ocat}
\title{Ordered categorical family}
\usage{
ocat(theta = NULL, link = "identity", R = NULL)
}
\description{
Ordered categorical distribution family for ordinal response data.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument.
}
\examples{
# ffc_gam(ordered_response ~ fts(x), data = data, family = ocat(), time = "time")
}
\seealso{
\code{\link[mgcv]{ocat}}
}
\keyword{internal}
