% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ffc_gam}
\alias{predict.ffc_gam}
\title{Predict from a fitted \pkg{ffc_gam} model}
\usage{
\method{predict}{ffc_gam}(object, newdata, type = "link", se.fit = FALSE, ...)
}
\arguments{
\item{object}{ a fitted \code{gam} object as produced by \code{gam()}.
                }

\item{newdata}{ A data frame or list containing the values of the model covariates at which predictions
  are required. If this is not provided then predictions corresponding to the
                original data are returned. If \code{newdata} is provided then
                it should contain all the variables needed for prediction: a
                warning is generated if not. See details for use with \code{link{linear.functional.terms}}. }

\item{type}{ When this has the value \code{"link"} (default) the linear predictor (possibly with
associated standard errors) is returned. When \code{type="terms"} each component of the 
linear predictor is returned seperately (possibly with standard errors): this includes 
parametric model components, followed by each smooth component, but excludes
any offset and any intercept. \code{type="iterms"} is the same, except that any standard errors 
returned for smooth components will include the uncertainty about the intercept/overall mean.  When 
\code{type="response"} predictions 
on the scale of the response are returned (possibly with approximate
standard errors). When \code{type="lpmatrix"} then a matrix is returned
which yields the values of the linear predictor (minus any offset) when
postmultiplied by the
parameter vector (in this case \code{se.fit} is ignored). The latter
option is most useful for getting variance estimates for quantities derived from
the model: for example integrated quantities, or derivatives of smooths. A
linear predictor matrix can also be used to implement approximate prediction
outside \code{R} (see example code, below). }

\item{se.fit}{ when this is TRUE (not default) standard error estimates are returned for each prediction.}

\item{...}{ignored}
}
\value{
If \code{type == "lpmatrix"} then a \code{matrix} is returned which
will give a vector of linear predictor values (minus any offest)
at the supplied covariate values, when applied to the model coefficient vector.
Otherwise, if \code{se.fit == TRUE} then a \code{2 }item \code{list} is returned with
items (both \code{arrays}) \code{fit} and \code{se.fit} containing predictions and
associated standard error estimates, otherwise an \code{array} of predictions is
returned. The dimensions of the returned \code{arrays} depends on whether
type is \code{"terms"} or not: if it is then the \code{array} is \code{2} dimensional with each
term in the linear predictor separate, otherwise the \code{array} is 1 dimensional and
contains the linear predictor/predicted values (or corresponding s.e.s).
The linear predictor returned termwise will not include the offset or the intercept.
}
\description{
Predict from a fitted \pkg{ffc_gam} model
}
\details{
This function returns predictions from models fitted with \code{\link[=ffc_gam]{ffc_gam()}}.
Data passed to \code{newdata} will first be correctly augmented to include any basis functions
whose coefficients were estimated as time-varying, so the user need only supply data
that includes variables that were used in the original \code{data} that was supplied to
\code{ffc_gam()}
}
\seealso{
\code{\link[=ffc_gam]{ffc_gam()}}, \code{\link[=fts]{fts()}}
}
\author{
Nicholas J Clark
}
