% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts_coefs.R
\name{print.fts_ts}
\alias{print.fts_ts}
\title{Print an fts_ts tibble}
\usage{
\method{print}{fts_ts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fts_ts}}

\item{...}{Ignored}
}
\description{
Print an fts_ts tibble
}
\author{
Nicholas J Clark
}
