% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qld_mortality.R
\docType{data}
\name{qld_mortality}
\alias{qld_mortality}
\title{Queensland mortality data}
\format{
A \code{data.frame} containing the following fields:
\describe{
\item{year}{integer, year of records}
\item{age}{integer, age at death, in years}
\item{sex}{factor differentiating the sexes}
\item{deaths}{integer, number of deaths recorded}
\item{population}{numeric population recorded at 30 June each year}
}
}
\source{
Australian Human Mortality Database
}
\usage{
qld_mortality
}
\description{
A dataset containing number of deaths, population at risk and
ages for residents of Queensland, Australia from 1979 to 2020
}
\details{
The age group 100 also includes people who died aged older than 100.
The data come from the Australian Human Mortality Database (\url{https://aushd.org}).
}
\keyword{datasets}
