% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{rTweedie}
\alias{rTweedie}
\title{Random Tweedie generation}
\usage{
rTweedie(mu, p = 1.5, phi = 1)
}
\description{
Generates random numbers from Tweedie distribution.
Utility function for Tweedie family.
}
\examples{
# rTweedie(n = 100, mu = 1, p = 1.5, phi = 1)
}
\seealso{
\code{\link[mgcv]{rTweedie}}, \code{\link[=tw]{tw()}}
}
