% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{s}
\alias{s}
\title{Smooth terms for use in GAMs}
\usage{
s(
  ...,
  k = -1,
  fx = FALSE,
  bs = "tp",
  m = NA,
  by = NA,
  xt = NULL,
  id = NULL,
  sp = NULL,
  pc = NULL
)
}
\description{
Constructs smooth terms for inclusion in GAM formulae. Use within
\code{\link[=ffc_gam]{ffc_gam()}} for standard smooths or \code{\link[=fts]{fts()}} for time-varying smooths.
}
\examples{
# Standard smooth in ffc_gam
# ffc_gam(y ~ s(x), data = data, time = "time")
}
\seealso{
\code{\link[mgcv]{s}}
}
