% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{scat}
\alias{scat}
\title{Scaled Student-t family}
\usage{
scat(theta = NULL, link = "identity", min.df = 3)
}
\description{
Scaled Student-t distribution family for robust regression.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument or Stan models.
}
\examples{
# ffc_gam(y ~ fts(x), data = data, family = scat(), time = "time")
}
\seealso{
\code{\link[mgcv]{scat}}
}
\keyword{internal}
