% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{te}
\alias{te}
\title{Tensor product smooth terms}
\usage{
te(
  ...,
  k = NA,
  bs = "cr",
  m = NA,
  d = NA,
  by = NA,
  fx = FALSE,
  np = TRUE,
  xt = NULL,
  id = NULL,
  sp = NULL,
  pc = NULL
)
}
\description{
Constructs tensor product smooth terms for multidimensional smoothing.
Use within \code{\link[=ffc_gam]{ffc_gam()}} formulae for interactions between variables.
}
\examples{
# Tensor product smooth in ffc_gam
# ffc_gam(y ~ te(x1, x2), data = data, time = "time")
}
\seealso{
\code{\link[mgcv]{te}}
}
