% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{tw}
\alias{tw}
\title{Tweedie family}
\usage{
tw(theta = NULL, link = "log", a = 1.01, b = 1.99)
}
\description{
Tweedie distribution family for continuous data with zeros.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument.
}
\examples{
# ffc_gam(continuous_with_zeros ~ fts(x), data = data, family = tw(), time = "time")
}
\seealso{
\code{\link[mgcv]{tw}}
}
