% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{twlss}
\alias{twlss}
\title{Tweedie location-scale-shape family}
\usage{
twlss(link = list("log", "identity", "identity"), a = 1.01, b = 1.99)
}
\description{
Extended Tweedie family allowing shape and scale parameters to vary.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument for complex Tweedie models.
}
\examples{
# ffc_gam(list(y ~ fts(x), ~ s(z), ~ s(w)), family = twlss(), data = data, time = "time")
}
\seealso{
\code{\link[mgcv]{twlss}}, \code{\link[=tw]{tw()}}
}
\keyword{internal}
