% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{ziP}
\alias{ziP}
\title{Zero-inflated Poisson family}
\usage{
ziP(theta = NULL, link = "identity", b = 0)
}
\description{
Zero-inflated Poisson distribution family for count data with excess zeros.
Use in \code{\link[=ffc_gam]{ffc_gam()}} \code{family} argument.
}
\examples{
# ffc_gam(zero_inflated_counts ~ fts(x), data = data, family = ziP(), time = "time")
}
\seealso{
\code{\link[mgcv]{ziP}}
}
\keyword{internal}
