% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffc-package.R
\docType{package}
\name{ffc-package}
\alias{ffc}
\alias{ffc-package}
\title{ffc: Functional ForeCasting}
\description{
Functional forecasting using Generalized Additive Models
}
\author{
\strong{Maintainer}: Nicholas J Clark \email{nicholas.j.clark1214@gmail.com} (\href{https://orcid.org/0000-0001-7131-3301}{ORCID})

}
\keyword{internal}
