% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.ffc_gam}
\alias{forecast.ffc_gam}
\title{Forecasting \code{ffc_gam} models}
\usage{
\method{forecast}{ffc_gam}(
  object,
  newdata,
  type = "response",
  n_draws = 25,
  n_sims = 50,
  model = "ARIMA",
  summary = TRUE,
  robust = TRUE,
  probs = c(0.025, 0.975),
  ...
)
}
\arguments{
\item{object}{An object of class \code{ffc_gam}. See \code{\link[=ffc_gam]{ffc_gam()}}}

\item{newdata}{\code{dataframe} or \code{list} of test data containing the
same variables that were included in the original \code{data} used to fit the
model. The covariate information in \code{newdata}, along with the temporal information
indexed by the \code{time} variable in the original call to \code{ffc_gam()}, will be used to
generate forecasts from the fitted model equations}

\item{type}{When this has the value \code{link}, the linear predictor is calculated
on the link scale. If \code{expected} is used, predictions reflect the expectation
of the response (the mean) but ignore uncertainty in the observation process.
When \code{response} is used (the default), the predictions take uncertainty in the observation
process into account to return predictions on the outcome scale}

\item{n_draws}{A positive \code{integer} specifying the number of basis coefficient
realisation paths to simulate from the fitted time-varying \code{fts()} splines}

\item{n_sims}{A positive \code{integer} specifying the number of simulation paths to compute
from each basis coefficient time series model when creating the basis coefficient
forecast distributions}

\item{model}{A \code{character} string representing a valid univariate model definition
from the \pkg{fable} package. Note that the chosen method must have an associated
\code{generate()} method in order to simulate forecast realisations. Valid models
currently include: \code{'ETS'}, \code{'ARIMA'}, \code{'AR'}, \code{'NAIVE'}, and \code{'NNETAR'}}

\item{summary}{Should summary statistics be returned instead of the raw values?
Default is \code{TRUE}}

\item{robust}{If \code{FALSE} (the default) the mean is used as the measure of
central tendency and the standard deviation as the measure of variability.
If \code{TRUE}, the median and the median absolute deviation (MAD) are applied instead.
Only used if \code{summary} is \code{TRUE}}

\item{probs}{The percentiles to be computed by the \code{quantile()} function.
Only used if \code{summary} is \code{TRUE}}

\item{...}{ignored}
}
\value{
Predicted values on the appropriate scale.
If \code{summary == FALSE}, the output is a matrix. If \code{summary == TRUE}, the output
is a tidy \code{tbl_df / data.frame}
}
\description{
Forecasting \code{ffc_gam} models
}
\details{
Computes forecast distributions from fitted \code{ffc_gam} objects
}
\seealso{
\code{\link[=ffc_gam]{ffc_gam()}}, \code{\link[=fts]{fts()}}, \code{\link[=forecast.fts_ts]{forecast.fts_ts()}}
}
\author{
Nicholas J Clark
}
