% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts_coefs.R
\name{forecast.fts_ts}
\alias{forecast.fts_ts}
\alias{forecast}
\title{Forecasting functional basis coefficients}
\usage{
\method{forecast}{fts_ts}(object, model = ETS(), h = 1, times = 10, ...)
}
\arguments{
\item{object}{An object of class \code{fts_ts} containing time-varying
basis function coefficients extracted from an \code{ffc_gam} object}

\item{model}{A valid model definition from the \pkg{fable} package
(i.e. \code{ETS()}, \code{ARIMA()} etc...)}

\item{h}{A positive \code{integer} specifying the length of the forecast
horizon}

\item{times}{A positive \code{integer} specifying the number of forecast
realisation paths to simulate from the fitted forecast \code{model}}

\item{...}{Ignored}
}
\value{
A \code{tsibble} object containing the forecast prediction (\code{.sim}) for each
replicate realisation (\code{.rep}) at each timestep in the forecast horizon \code{h}
}
\description{
Forecasting functional basis coefficients
}
\details{
Basis function coefficient time series will be used as input
to the specified \code{model} to train a forecasting model that will then be
extrapolated \code{h} timesteps into the future. A total of \code{times} forecast realisations
will be returned for each basis coefficient
}
\author{
Nicholas J Clark
}
