% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.fts_ts}
\alias{forecast.fts_ts}
\title{Forecasting functional basis coefficients}
\usage{
\method{forecast}{fts_ts}(object, model = "ARIMA", h = 1, times = 25, stationary = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{fts_ts} containing time-varying
basis function coefficients extracted from an \code{ffc_gam} object}

\item{model}{A \code{character} string representing a valid univariate model definition
from the \pkg{fable} package or one of the built-in Bayesian dynamic
factor models. Note that if a \pkg{fable} model is used,
the chosen method must have an associated
\code{generate()} method in order to simulate forecast realisations. Valid models
currently include: \code{'ARDF'}, \code{'GPDF'}, '\code{VARDF},
\code{'ETS'}, \code{'ARIMA'}, \code{'AR'}, \code{'RW'}, \code{'NAIVE'}, and \code{'NNETAR'}}

\item{h}{A positive \code{integer} specifying the length of the forecast
horizon}

\item{times}{A positive \code{integer} specifying the number of forecast
realisation paths to simulate from the fitted forecast \code{model}}

\item{stationary}{If \code{TRUE}, the fitted time series models are
constrained to be stationary. Default is \code{FALSE}. This option
only works when \code{model == 'ARIMA'}}

\item{...}{Other arguments to pass to the Stan dynamic factor models
(i.e. the (V)AR order \code{lag = ...} or the number of factors \code{K = ...})}
}
\value{
A \code{tsibble} object containing the forecast prediction (\code{.sim}) for each
replicate realisation (\code{.rep}) at each timestep in the forecast horizon \code{h}
}
\description{
Forecasting functional basis coefficients
}
\details{
Basis function coefficient time series will be used as input
to the specified \code{model} to train a forecasting model that will then be
extrapolated \code{h} timesteps into the future. A total of \code{times} forecast realisations
will be returned for each basis coefficient
}
\examples{
\donttest{
# Extract coefficients and generate forecasts
mod <- ffc_gam(
  deaths ~ offset(log(population)) + sex + 
    fts(age, k = 8, bs = "cr", time_k = 10),
  time = "year", 
  data = qld_mortality,
  family = poisson(),
  engine = "bam"
)
coefs <- fts_coefs(mod, summary = FALSE, times = 5)

# Generate ARDF forecasts
fc_ardf <- forecast(coefs, h = 3, model = "ARDF", K = 2)

# Plot the forecasts
autoplot(fc_ardf)
}
}
\seealso{
\code{\link[=predict]{predict()}}, \code{\link[=fts]{fts()}}
}
\author{
Nicholas J Clark
}
