% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts_coefs.R
\name{fts_coefs.ffc_gam}
\alias{fts_coefs.ffc_gam}
\alias{fts_coefs}
\title{Extract time-varying basis coefficients}
\usage{
fts_coefs(object, ...)

\method{fts_coefs}{ffc_gam}(object, summary = TRUE, n_samples = 25, ...)
}
\arguments{
\item{object}{\code{list} object of class \code{ffc_gam}. See \code{\link[=ffc_gam]{ffc_gam()}}}

\item{...}{Ignored}

\item{summary}{\code{Logical}. Should summary statistics of the coefficient
time series be returned instead of realized curves? Default is \code{TRUE}. If
\code{FALSE}, replicate realisations of each basis coefficient time series will
be returned}

\item{n_samples}{A positive \code{integer} specifying the number of time series
realisation paths to simulate from the fitted model. Ignored if
\code{summary = FALSE}}
}
\value{
A \code{fts_ts} object containing the point estimates and their standard errors
for basis function coefficients
}
\description{
Extract time-varying basis coefficients a fitted \pkg{ffc_gam} model so they can
be visualized and / or forecasted ahead
}
\details{
This function creates a \code{tidy} time series of basis function coefficients
from all \code{fts()} terms that were supplied to the original model
}
\seealso{
\code{\link[=ffc_gam]{ffc_gam()}}, \code{\link[=fts]{fts()}}
}
\author{
Nicholas J Clark
}
