% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts_coefs.R
\name{fts_coefs.ffc_gam}
\alias{fts_coefs.ffc_gam}
\alias{fts_coefs}
\title{Extract time-varying basis coefficients}
\usage{
fts_coefs(object, ...)

\method{fts_coefs}{ffc_gam}(object, ...)
}
\arguments{
\item{object}{\code{list} object of class \code{ffc_gam}. See \code{\link[=ffc_gam]{ffc_gam()}}}

\item{...}{Ignored}
}
\value{
A \code{fts_ts} object containing the point estimates and their standard errors
for basis function coefficients
}
\description{
Extract time-varying basis coefficients a fitted \pkg{ffc_gam} model so they can
be visualized and / or forecasted ahead
}
\details{
This function creates a \code{tidy} time series of basis function coefficients
from all \code{fts()} terms that were supplied to the original model
}
\seealso{
\code{\link[=ffc_gam]{ffc_gam()}}, \code{\link[=fts]{fts()}}
}
\author{
Nicholas J Clark
}
