% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts_coefs.R
\name{fts_coefs.ffc_gam}
\alias{fts_coefs.ffc_gam}
\alias{fts_coefs}
\title{Extract time-varying basis coefficients}
\usage{
fts_coefs(object, ...)

\method{fts_coefs}{ffc_gam}(object, se.fit = TRUE, ...)
}
\arguments{
\item{object}{\code{list} object of class \code{ffc_gam}. See \code{\link[=ffc_gam]{ffc_gam()}}}

\item{...}{Ignored}

\item{se.fit}{\code{logical} indicating whether standard errors of time-varying
basis coefficients should also be returned. Default is \code{TRUE}}
}
\value{
If \code{type == "lpmatrix"} then a \code{matrix} is returned which
will give a vector of linear predictor values (minus any offest)
at the supplied covariate values, when applied to the model coefficient vector.
Otherwise, if \code{se.fit == TRUE} then a \code{2 }item \code{list} is returned with
items (both \code{arrays}) \code{fit} and \code{se.fit} containing predictions and
associated standard error estimates, otherwise an \code{array} of predictions is
returned. The dimensions of the returned \code{arrays} depends on whether
type is \code{"terms"} or not: if it is then the \code{array} is \code{2} dimensional with each
term in the linear predictor separate, otherwise the \code{array} is 1 dimensional and
contains the linear predictor/predicted values (or corresponding s.e.s).
The linear predictor returned termwise will not include the offset or the intercept.
}
\description{
Extract time-varying basis coefficients a fitted \pkg{ffc_gam} model so they can
be visualized and / or forecasted ahead
}
\details{
This function returns predictions from models fitted with \code{\link[=ffc_gam]{ffc_gam()}}.
Data passed to \code{newdata} will first be correctly augmented to include any basis functions
whose coefficients were estimated as time-varying, so the user need only supply data
that includes variables that were used in the original \code{data} that was supplied to
\code{ffc_gam()}
}
\seealso{
\code{\link[=ffc_gam]{ffc_gam()}}, \code{\link[=fts]{fts()}}
}
\author{
Nicholas J Clark
}
