% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{s}
\alias{te}
\alias{gam}
\alias{bam}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mgcv}{\code{\link[mgcv]{bam}}, \code{\link[mgcv]{gam}}, \code{\link[mgcv]{s}}, \code{\link[mgcv]{te}}}
}}

