% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{s}
\alias{te}
\alias{gam}
\alias{bam}
\alias{nb}
\alias{negbin}
\alias{betar}
\alias{cnorm}
\alias{ocat}
\alias{scat}
\alias{ziP}
\alias{multinom}
\alias{tw}
\alias{ldTweedie}
\alias{rTweedie}
\alias{twlss}
\alias{autoplot}
\alias{forecast}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{forecast}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{mgcv}{\code{\link[mgcv]{bam}}, \code{\link[mgcv:Beta]{betar}}, \code{\link[mgcv]{cnorm}}, \code{\link[mgcv]{gam}}, \code{\link[mgcv]{ldTweedie}}, \code{\link[mgcv]{multinom}}, \code{\link[mgcv:negbin]{nb}}, \code{\link[mgcv]{negbin}}, \code{\link[mgcv]{ocat}}, \code{\link[mgcv]{rTweedie}}, \code{\link[mgcv]{s}}, \code{\link[mgcv]{scat}}, \code{\link[mgcv]{te}}, \code{\link[mgcv:Tweedie]{tw}}, \code{\link[mgcv]{twlss}}, \code{\link[mgcv]{ziP}}}
}}

